Fixes to run under Python 3 (and thus under ESX 6.5 GA)  (#756)
Looks like we do not need bilingual "six" module and just needed a few simple fixes to work  in Py3. This commit adds these fixes (note: this commit also removes vol_test.py - it was never really used and was always skipped) 

* Fixes to run with Python3

+++ esx_service/utils/cli_table.py
@@ -102,7 +102,7 @@ def shrink_to_fit(column_sizes, terminal_width):
 
         if gap * num_largest_columns > to_remove:
             # We can resize in this step and we are done
-            to_remove_per_column = to_remove / num_largest_columns
+            to_remove_per_column = int(to_remove / num_largest_columns)
             remainder = to_remove % num_largest_columns
             for i in range(num_largest_columns):
                 sorted_sizes[i] = largest - to_remove_per_column
